Refactor styles: move components from Boxstyles to Messagestyles and update imports in MessageForm and MessageItem
